+++ community-modules/client-side-row-model/src/clientSideNodeManager/clientSideNodeManager.ts
@@ -15,20 +15,37 @@ import {
     _warnOnce,
 } from '@ag-grid-community/core';
 
-import type { ClientSideNodeManagerRootNode, ClientSideNodeManagerRowNode } from './abstractClientSideNodeManager';
-import { AbstractClientSideNodeManager } from './abstractClientSideNodeManager';
+import type {
+    ClientSideNodeManagerRootNode,
+    ClientSideNodeManagerRowNode,
+} from '../clientSideNodeManager/abstractClientSideNodeManager';
+import { AbstractClientSideNodeManager } from '../clientSideNodeManager/abstractClientSideNodeManager';
+import { makeFieldPathGetter } from './fieldAccess';
+import type { DataFieldGetter } from './fieldAccess';
 
 const TOP_LEVEL = 0;
 
-export class ClientSideNodeManager<TData>
+export class ClientSideTreeNodeManager<TData>
     extends AbstractClientSideNodeManager<TData>
     implements IClientSideNodeManager<TData>, NamedBean
 {
-    beanName = 'clientSideNodeManager' as const;
+    beanName = 'clientSideTreeNodeManager' as const;
+
+    private allNodesMap: { [id: string]: RowNode } = {};
+
+    private childrenGetter: DataFieldGetter | null = null;
 
     // when user is provide the id's, we also keep a map of ids to row nodes for convenience
-    private allNodesMap: { [id: string]: RowNode } = {};
     private nextId = 0;
+
+    public override initRootNode(rootRowNode: RowNode<TData>): void {
+        const childrenField = this.gos.get('treeDataChildrenField');
+        if (this.childrenGetter?.path !== childrenField) {
+            this.childrenGetter = makeFieldPathGetter(childrenField);
+        }
+
+        super.initRootNode(rootRowNode);
+    }
 
     public getRowNode(id: string): RowNode | undefined {
         return this.allNodesMap[id];
